feat(xtal): store density maps in section chunks with 64-bit indexing
DensityMap kept its 8-bit samples in one contiguous new[] block indexed
with int arithmetic. A 1300^3 cryo-EM map overflows the int voxel count,
and inside the Electron process PartitionAlloc refuses single allocations
above about 2 GiB, so large maps could not be opened at all.

- New qlib::ChunkedArray3D<T>: the volume is split along the section axis
  into chunks of at most 8 MiB (one allocation each), element order inside
  a chunk equal to Array3D (column fastest) so slices and rows stay
  contiguous, size_t indexing, and a Layout that can be computed without
  allocating. DensityMap now holds a ChunkedArray3D<quint8> value member
  (the delete[]-of-scalar-new bug on reload disappears with the pointer).
- DensityMap gains a streaming fill API for readers: beginByteMap(dims,
  quantization) / sliceBytes(k) / endByteMap(stats). setMapFloatArray and
  setMapByteArray are built on it; the statistics arithmetic is unchanged
  so the MapSurfPin goldens hold. The Brix and QDF readers write their
  sections straight into the map instead of a whole-map temporary, and
  the QDF writer emits each section as one fixed-record block through the
  new QdfOutStream::writeFxRecords() instead of three calls per voxel.
- Array3D keeps its contiguous data() contract (GL texture uploads depend
  on it) and only gets size_t sizes / indices. The other int voxel counts
  (CCP4, Xplor, QDF, ElePotMap, MapFFT, readFxRecords) become size_t.
- Histogram: ScalarObject::getBaseHistogram() lets an implementation
  supply a lossless base histogram; DensityMap returns the 256-bin byte
  histogram, built chunk-parallel on first use, instead of scanning every
  voxel through the virtual atFloat() into an rmsd/1000-wide table. The
  generic path uses 64-bit counts and caps the bin count. The cache is
  dropped on reload (it used to survive a re-read). getHistogramJSON()
  keeps its JSON contract. getLevelAtTopFraction(frac) (scriptable) gives
  the level enclosing the top fraction of the grid points from the same
  histogram.
- Remove the orphaned DenRealMap.cpp (never built, no header).

Tests: chunk layout / indexing equivalence with Array3D / copy-move
(test_qlib), histogram rebinning equivalence, reload invalidation and the
top-fraction level (test_xtal).

+++ src/modules/surface/ElePotMap.cpp
@@ -59,8 +59,8 @@ bool ElePotMap::setMapFloatArray(const float *array,
   rhomean=0.0, sqmean=0.0,
   rhodev=0.0;
 
-  const int ntotal = ncol*nrow*nsect;
-  for (int i=0; i<ntotal; i++) {
+  const size_t ntotal = size_t(ncol)*size_t(nrow)*size_t(nsect);
+  for (size_t i=0; i<ntotal; i++) {
     double rho = (double)array[i];
     rhomean += rho/float(ntotal);
     sqmean += rho*rho/float(ntotal);
